# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment System API",
    "description": "A comprehensive API for managing assessments, submissions, and reviews in a sustainability assessment system.\n\n## Authentication\nAll endpoints require JWT authentication. Include the token in the Authorization header:\n`Authorization: Bearer <your-jwt-token>`\n\n## Versioning\nThe API supports response versioning to track changes and resolve conflicts.\n",
    "contact": {
      "name": "Assessment System Team",
      "email": "support@assessment-system.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/admin/reports": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all reports across organizations",
        "description": "Get all reports for all organizations (DGRV admin only)",
        "operationId": "list_all_reports",
        "responses": {
          "200": {
            "description": "List of all organization reports",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportListResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/questions": {
      "get": {
        "tags": ["Questions"],
        "summary": "List questions",
        "description": "Get all questions",
        "responses": {
          "200": {
            "description": "List of questions with their revisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListWithRevisionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Questions"],
        "summary": "Create question",
        "description": "Create a new question (Admin only)",
        "operationId": "create_question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/questions/{question_id}": {
      "get": {
        "tags": ["Questions"],
        "summary": "Get question",
        "description": "Get specific question with all revisions or a specific revision",
        "operationId": "get_question",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revision_id",
            "in": "query",
            "description": "Specific revision ID to retrieve (optional)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Language for content",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionWithRevisionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Questions"],
        "summary": "Update question",
        "description": "Update a question (Admin only). Include question_revision_id for optimistic locking.",
        "operationId": "update_question",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "question_revision_id",
            "in": "query",
            "description": "Current revision ID for optimistic locking",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Conflict - revision mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/category-catalog": {
      "get": {
        "tags": ["Category Catalog"],
        "summary": "List category catalogs",
        "description": "Get all active category catalogs",
        "operationId": "get_category_catalogs",
        "responses": {
          "200": {
            "description": "List of active category catalogs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Category Catalog"],
        "summary": "Create category catalog",
        "description": "Create a new category catalog entry (Admin only)",
        "operationId": "create_category_catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Category catalog created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/assessments": {
      "get": {
        "tags": ["Assessments"],
        "summary": "List assessments (org-scoped)",
        "description": "Get all draft and submitted assessments for the current user's organization. Accessible to both org_admin and Org_User.",
        "operationId": "list_assessments",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "Language for content",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by assessment status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["draft", "submitted"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of assessments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Assessments"],
        "summary": "Create assessment (org-scoped)",
        "description": "Create a new draft assessment for the current user's organization. Only accessible to org_admin.",
        "operationId": "create_assessment",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "language": {
                    "type": "string",
                    "default": "en",
                    "description": "Language for the assessment"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assessment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/assessments/{assessment_id}": {
      "get": {
        "tags": ["Assessments"],
        "summary": "Get assessment",
        "description": "Get a specific draft assessment",
        "operationId": "get_assessment",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assessment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDetailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Assessments"],
        "summary": "Update assessment",
        "description": "Update a draft assessment",
        "operationId": "update_assessment",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Assessment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Assessments"],
        "summary": "Delete assessment",
        "description": "Delete a draft assessment",
        "operationId": "delete_assessment",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Assessment deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/assessments/{assessment_id}/submit": {
      "post": {
        "tags": ["Assessments"],
        "summary": "Approve and submit assessment",
        "description": "Approve and submit a draft assessment (admin only - creates immutable submission)",
        "operationId": "submit_assessment",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Assessment approved and submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/assessments/{assessment_id}/responses": {
      "get": {
        "tags": ["Responses"],
        "summary": "List responses",
        "description": "Get all responses for a draft assessment",
        "operationId": "list_responses",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "latest_only",
            "in": "query",
            "description": "Return only latest version of each response",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of responses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Responses"],
        "summary": "Create batch of responses",
        "description": "Create or update a batch of responses to questions for a given assessment.",
        "operationId": "create_batch_responses",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateResponseRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Responses created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/assessments/{assessment_id}/responses/{response_id}": {
      "get": {
        "tags": ["Responses"],
        "summary": "Get response",
        "description": "Get a specific response",
        "operationId": "get_response",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "response_id",
            "in": "path",
            "description": "Response ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDetailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Responses"],
        "summary": "Update response",
        "description": "Update a response (creates new version)",
        "operationId": "update_response",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "response_id",
            "in": "path",
            "description": "Response ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Responses"],
        "summary": "Delete response",
        "description": "Delete a response",
        "operationId": "delete_response",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "response_id",
            "in": "path",
            "description": "Response ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Response deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/submissions": {
      "get": {
        "tags": ["Submissions"],
        "summary": "List user submissions",
        "description": "Get user's submitted assessments",
        "operationId": "list_user_submissions",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by review status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending_review",
                "under_review",
                "approved",
                "rejected",
                "revision_requested"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user submissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/admin/action-plans": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all action plans",
        "description": "Get all action plans for all organizations (DGRV admin only)",
        "operationId": "list_all_action_plans",
        "responses": {
          "200": {
            "description": "List of all organization action plans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlanListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminReportListResponse": {
        "type": "object",
        "properties": {
          "reports": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminReport" }
          }
        },
        "required": ["reports"]
      },
...
      "UpdateAssessmentRequest": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}